CO3109-Shop-Gateway-Python/requirements.txt
flask
opencv-python
numpy
paho-mqtt
python-dotenv
requests
